Moved 'get tag requirements' api call to PreservationTab component.
