+++ src/modules/Preservation/views/ScopeOverview/TagFlyout/PreservationTab/PreservationTab.tsx
@@ -1,37 +1,59 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 
 import { Container, TagDetailsContainer, Details, GridFirstRow, GridSecondRow, RemarkContainer } from './PreservationTab.style';
 import { TextField, Typography } from '@equinor/eds-core-react';
 import { TagDetails, TagRequirement } from './../types';
 import Requirements from './Requirements';
+import Spinner from '../../../../../../components/Spinner';
+import { usePreservationContext } from '../../../../context/PreservationContext';
+import { showSnackbarNotification } from './../../../../../../core/services/NotificationService';
 
 interface PreservationTabProps {
-    details: TagDetails | undefined;
-    requirements: TagRequirement[] | undefined;
+    tagId: number | null;
+    tagDetails: TagDetails | undefined;
 }
 
 const PreservationTab = ({
-    details,
-    requirements
+    tagId,
+    tagDetails
 }: PreservationTabProps): JSX.Element => {
+    const [tagRequirements, setTagRequirements] = useState<TagRequirement[]>();
+    const { apiClient } = usePreservationContext();
 
-    if (details === undefined) {
-        return <div style={{margin: 'calc(var(--grid-unit) * 2)'}}>Missing data</div>;
+    const getTagRequirements = async (id: number): Promise<void> => {
+        try {            
+            const tagRequirements = await apiClient.getTagRequirements(id);
+            setTagRequirements(tagRequirements);
+        }
+        catch (error) {
+            console.error(`Get TagRequirements failed: ${error.message}`);
+            showSnackbarNotification(error.message, 5000);
+        }
+    };
+
+    useEffect(() => {
+        if (tagId !== null) {
+            getTagRequirements(tagId);
+        }
+    }, [tagId]);    
+
+    if (tagDetails === undefined) {
+        return <div style={{margin: 'calc(var(--grid-unit) * 5) auto'}}><Spinner medium /></div>;
     }
 
     return (
         <Container>
             <TagDetailsContainer>
                 <Details>
-                    <Typography variant='h6'>{details.description}</Typography>
+                    <Typography variant='h6'>{tagDetails.description}</Typography>
                     <div style={{marginTop: 'calc(var(--grid-unit) * 2)'}}>
                         <GridFirstRow>
                             <Typography variant='caption' style={{gridColumn: '1', gridRow: '1'}}>Journey</Typography>
                             <Typography variant='caption' style={{gridColumn: '2', gridRow: '1'}}>Mode</Typography>
                             <Typography variant='caption' style={{gridColumn: '3', gridRow: '1'}}>Resp.</Typography>
-                            <Typography variant='body_long' style={{gridColumn: '1', gridRow: '2'}}>{details.journeyTitle}</Typography>
-                            <Typography variant='body_long' style={{gridColumn: '2', gridRow: '2'}}>{details.mode}</Typography>
-                            <Typography variant='body_long' style={{gridColumn: '3', gridRow: '2'}}>{details.responsibleName}</Typography>
+                            <Typography variant='body_long' style={{gridColumn: '1', gridRow: '2'}}>{tagDetails.journeyTitle}</Typography>
+                            <Typography variant='body_long' style={{gridColumn: '2', gridRow: '2'}}>{tagDetails.mode}</Typography>
+                            <Typography variant='body_long' style={{gridColumn: '3', gridRow: '2'}}>{tagDetails.responsibleName}</Typography>
                         </GridFirstRow>
                     </div>
                     <div style={{marginTop: 'var(--grid-unit)'}}>
@@ -40,10 +62,10 @@ const PreservationTab = ({
                             <Typography variant='caption' style={{gridColumn: '2', gridRow: '1'}}>MC pkg</Typography>
                             <Typography variant='caption' style={{gridColumn: '3', gridRow: '1'}}>PO</Typography>
                             <Typography variant='caption' style={{gridColumn: '4', gridRow: '1'}}>Area</Typography>
-                            <Typography variant='body_short' style={{gridColumn: '1', gridRow: '2'}}>{details.commPkgNo}</Typography>
-                            <Typography variant='body_short' style={{gridColumn: '2', gridRow: '2'}}>{details.mcPkgNo}</Typography>
-                            <Typography variant='body_short' style={{gridColumn: '3', gridRow: '2'}}>{details.purchaseOrderNo}</Typography>
-                            <Typography variant='body_short' style={{gridColumn: '4', gridRow: '2'}}>{details.areaCode}</Typography>
+                            <Typography variant='body_short' style={{gridColumn: '1', gridRow: '2'}}>{tagDetails.commPkgNo}</Typography>
+                            <Typography variant='body_short' style={{gridColumn: '2', gridRow: '2'}}>{tagDetails.mcPkgNo}</Typography>
+                            <Typography variant='body_short' style={{gridColumn: '3', gridRow: '2'}}>{tagDetails.purchaseOrderNo}</Typography>
+                            <Typography variant='body_short' style={{gridColumn: '4', gridRow: '2'}}>{tagDetails.areaCode}</Typography>
                         </GridSecondRow>
                     </div>               
                 </Details>                
@@ -51,7 +73,7 @@ const PreservationTab = ({
             <RemarkContainer>
                 <TextField id='remark' label='Remark' disabled />
             </RemarkContainer>
-            <Requirements requirements={requirements} />
+            <Requirements requirements={tagRequirements} />
         </Container>
     );
 };
