+++ src/modules/Preservation/views/ScopeOverview/TagFlyout/TagFlyout.tsx
@@ -4,10 +4,9 @@ import { Container, Header, Tabs, StatusLabel, HeaderActions } from './TagFlyout
 import PreservationTab from './PreservationTab/PreservationTab';
 import CloseIcon from '@material-ui/icons/Close';
 import { Button } from '@equinor/eds-core-react';
-import Spinner from '../../../../../components/Spinner';
 import { usePreservationContext } from '../../../context/PreservationContext';
 import { showSnackbarNotification } from './../../../../../core/services/NotificationService';
-import { TagDetails, TagRequirement } from './types';
+import { TagDetails } from './types';
 
 interface TagFlyoutProps {
     close: () => void;
@@ -19,33 +18,16 @@ const TagFlyout = ({
     tagId
 }: TagFlyoutProps): JSX.Element => {
     const [activeTab, setActiveTab] = useState<string>('preservation');
-    const [isLoading, setIsLoading] = useState<boolean>();
     const [tagDetails, setTagDetails] = useState<TagDetails>();
-    const [tagRequirements, setTagRequirements] = useState<TagRequirement[]>();
     const { apiClient } = usePreservationContext();
 
     const getTagDetails = async (id: number): Promise<void> => {
-        try {            
-            setIsLoading(true);
+        try {
             const tagDetails = await apiClient.getTagDetails(id);
             setTagDetails(tagDetails);
         }
         catch (error) {
             console.error(`Get TagDetails failed: ${error.message}`);
-            showSnackbarNotification(error.message, 5000);
-        }
-        finally {
-            setIsLoading(false);
-        }
-    };
-
-    const getTagRequirements = async (id: number): Promise<void> => {
-        try {            
-            const tagRequirements = await apiClient.getTagRequirements(id);
-            setTagRequirements(tagRequirements);
-        }
-        catch (error) {
-            console.error(`Get TagRequirements failed: ${error.message}`);
             showSnackbarNotification(error.message, 5000);
         }
     };
@@ -56,22 +38,10 @@ const TagFlyout = ({
         }
     }, [tagId]);
 
-    useEffect(() => {
-        if (tagId !== null) {
-            getTagRequirements(tagId);
-        }
-    }, [tagId]);    
-
     const getTabContent = (): JSX.Element => {
-        if (isLoading) {
-            return (
-                <div style={{margin: 'calc(var(--grid-unit) * 5) auto'}}><Spinner medium /></div>
-            );
-        }
-
         switch (activeTab) {
             case 'preservation':
-                return <PreservationTab details={tagDetails} requirements={tagRequirements} />;
+                return <PreservationTab tagId={tagId} tagDetails={tagDetails} />;
             case 'actions':
                 return <div></div>;
             case 'attachments':
